backend/requirements.txt
faiss-cpu
langchain
langchain-community
langchain-core
moviepy==1.0.3
newspaper3k
numpy
ollama
pandas
PyPDF2
python-docx
fastapi
uvicorn
python-multipart
translate
urllib3
youtube-transcript-api
googletrans
pyttsx3
httpx
SpeechRecognition
pydub
lxml[html_clean]
lxml_html_clean
pyaudio
itsdangerous
whisper